+++ app/components/UI/Earn/__mocks__/earnMockData.ts
