+++ app/components/UI/Earn/utils/token/token.test.ts
@@ -8,7 +8,7 @@ import { TokenI } from '../../../Tokens/types';
 import {
   MOCK_ACCOUNT_MULTI_CHAIN_TOKENS,
   MOCK_SUPPORTED_EARN_TOKENS_NO_FIAT_BALANCE,
-} from '../../../Stake/__mocks__/mockData';
+} from '../../../Stake/__mocks__/stakeMockData';
 import { CHAIN_IDS } from '@metamask/transaction-controller';
 import {
   createMockToken,
