+++ app/components/UI/Stake/__mocks__/stakeMockData.ts
