+++ app/components/UI/Stake/components/StakingConfirmation/ConfirmationFooter/ConfirmationFooter.test.tsx
@@ -5,7 +5,7 @@ import { ConfirmationFooterProps } from './ConfirmationFooter.types';
 import { createMockAccountsControllerState } from '../../../../../../util/test/accountsControllerTestUtils';
 import { backgroundState } from '../../../../../../util/test/initial-root-state';
 import { FooterButtonGroupActions } from './FooterButtonGroup/FooterButtonGroup.types';
-import { MOCK_POOL_STAKING_SDK } from '../../../__mocks__/mockData';
+import { MOCK_POOL_STAKING_SDK } from '../../../__mocks__/stakeMockData';
 
 const MOCK_ADDRESS_1 = '0x0';
 const MOCK_ADDRESS_2 = '0x1';
