+++ app/components/UI/Stake/components/StakingConfirmation/ConfirmationFooter/FooterButtonGroup/FooterButtonGroup.test.tsx
@@ -9,7 +9,7 @@ import {
 } from './FooterButtonGroup.types';
 import { createMockAccountsControllerState } from '../../../../../../../util/test/accountsControllerTestUtils';
 import { backgroundState } from '../../../../../../../util/test/initial-root-state';
-import { MOCK_POOL_STAKING_SDK } from '../../../../__mocks__/mockData';
+import { MOCK_POOL_STAKING_SDK } from '../../../../__mocks__/stakeMockData';
 
 const MOCK_ADDRESS_1 = '0x0';
